requirements-portable.txt
# WakeFusion portable runtime requirements
# Device-side ASR/TTS and training dependencies are intentionally excluded.

numpy>=1.24.0
sounddevice>=0.4.6
pyaudio>=0.2.13

openwakeword>=0.6.0
sherpa-onnx>=1.10.0
pvporcupine>=1.9.0
nemo-toolkit[asr]>=1.14.0
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
webrtcvad>=2.0.10

opencv-python>=4.9.0
mediapipe>=0.10.14
pyorbbecsdk>=0.1.0

pydantic>=2.0
aiohttp>=3.9.0
websockets>=11.0
pyyaml>=6.0
pyzmq>=25.0.0
psutil>=5.9.0
colorlog>=6.7.0
